Improved error handling #2
Both read and write did not handle errors properly. In particular,
errors happening in either the functions passed into thread_io by the
client caused unexpected panics or hangs.

Added a bunch of tests to handle the various conditions, and improved
some existing tests. Ensured as much as possible that the most
informative errors is returned in all cases.